get experiments running with cgroups
 * create some scripts to
   * create a cgroup
   * configure it
   * run programs in it

 * I had to fix the ycsb build, since we're now requiring
   that we build veribetrfs with libc++ for perf reasons.
   * RocksDB is still built with the *system default* standard
     library, which is gcc's standard library on most linux systems.
   * This means that the ycsb lib needs to build to be compatible
     with either library, and the ycsb binary needs to be split into
     RocksYcsb and VeribetrfsYcsb.
   * This is annoying, but it does mean we can just pick the best
     one for each independently.

TODO Should probably benchmark both for rocks to see which
one is faster for rocks and use that one to be fair.

TODO If we make our own shared_ptr library, we wouldn't need to use
libc++ just for that optimization, and maybe it would turn out
libstdc++ is actually faster.

TODO we can compare not just libc++ vs libstdc++ but also gcc vs clang.