nr: introduce tracking of already read bits
in the end when we bump the local tail, we must have read
everything between the old and new local tail. this implies
that all entries in between must have been alive. This
change introduces a "current" pointer in the Reader state
that tracks which entry is the next to be read.

Signed-off-by: Reto Achermann <achreto@gmail.com>