try_combine if update isn't ready when checking for response.
Rust included this, but it was missed in porting to Dafny.
Without this try_combine some threads may (1) put an update in their
local operation buffer, (2) check and see that a combiner exists,
(3) the combiner is far enough along that it doesn't see the
thread's operation, (4) all other threads only issue read ops
so they never combine again, leading to permanent non-progress
on the thread awaiting its update result.

This patch ports the call to try_combine into that spin loop along
with the voodoo constant delay before attempting combining.