CoordinationSystemRefinement recover step invariant
Woooo!!! Finally got it. Key takeaway for me rn is that when doing
algebra on paper there's many properties you learn as a kid that you
take for granted that you have to remember that verifier doesn't
necessarily remember (or can't know for free).

In this case it's that if you have some value (a + b + c), and you
know that (b + c) = d, you don't immediately get that (a + b + c) =
(a + d) unless you know that addition is associative!!!

So crazy to think that these things underpin all of it.

TODO:
- Clean up assertions in the recover step proof