CoordinationSystem Put Step Refines!!!
Holy macaroli, came down to setting up a bunch of extensionality functions
and ALSO at the very core after unwrapping 3 layers of MapSpec tomfoolery
it was a `show` lemma that was necessary?! HUH? WHY!?

AAAAAAHHHH

Seriously though. If I'm trying to prove "there exists some step such
that this thing happens". Then I show "this thing happens with these
args", why does it take a lemma for the verifier to believe that
there exists some step such that stuff happens?